refactor: move User, User_Auth, Provider to common
