+++ src/main/java/com/bovo/Bovo/common/Participation.java
@@ -1,5 +1,4 @@
 package com.bovo.Bovo.common;
-import com.bovo.Bovo.common.domain.Users;
 import jakarta.persistence.*;
 import lombok.*;
