+++ src/main/java/com/bovo/Bovo/common/Provider.java
@@ -1,4 +1,4 @@
-package com.bovo.Bovo.common.domain;
+package com.bovo.Bovo.common;
 
 public enum Provider {
     LOCAL, KAKAO
