+++ src/main/java/com/bovo/Bovo/common/User_Auth.java
@@ -1,4 +1,4 @@
-package com.bovo.Bovo.common.domain;
+package com.bovo.Bovo.common;
 
 import jakarta.persistence.*;
 import lombok.*;
